RestApiClaimProviderTest
